// src/services/ridivi/interfaces/get-iban-account-response.ts
export interface GetIbanAccountResponse {
    currency: string;
    accountNumber: string;
    description: string;
    identification: string;
    idType: number;
    nameOwner: string;
    accountState: string;
    entity: string;
    entityName: string;
    countryCode: string;
    countryname: string;
}